Revert "Fix showing tracks of selected artist on 64-bit systems"
This reverts commit 3732944d096943125e4a51f753644a00726acc68.

+++ plugins/sorttab_display/normal_sorttab_page.c
@@ -66,7 +66,7 @@ struct _NormalSortTabPagePrivate {
      * Handler id of the idle callback when the selection
      * changed callback is executed
      */
-    gulong selection_changed_idle_id;
+    guint selection_changed_idle_id;
 
     /* name of entry last selected */
     GList *last_selection;
